Add Travis CI build support and other updates (#4)
Disabled Docker Hub autobuild for amd64 and implemented Travis CI to build and push Docker images for amd64 and armhf. Builds happen automatically on pushes to master and will tag builds with amd64, armhf, amd64-VERSION, and armhf-VERSION, and additionally update the fat manifest for marthoc/deconz:latest for multiarch support. This will make bumping deCONZ versions much more efficient going forward and will allow users the ability to roll-back to previous deCONZ versions. Rebased the armhf variant to resin/rpi-raspbian as part of the Travis CI rollout.

Updated Dockerfiles and README.md to remove requirement of running the container as privileged.

Added downstream Hass.io add-on images to this repository to participate in the Travis CI build pipeline.