Remove support for building Hass.io Add-on images
Now that the official Hass.io deCONZ Add-on has achieved (and even surpassed) feature parity with the marthoc deCONZ Add-on, the marthoc Add-on should be considered discontinued. This repo autobuilt the Hass.io Add-on images at the same time as the regular Docker images, but this is no longer required, so all support for building the Hass.io images is removed by this commit.