Fix Conbee 2 by upgrading OS/Window Manager
Short:

    * Upgrade to Ubuntu 20.04
    * Switch from wmii to i3-wm

Long:

I have a VM with USB Passthrough running on ESXi 7.0, however the marthoc/deconz image was not running properly and caused the docker process to hang up and lose connection with the Conbee 2 USB stick repeatedly.

This resulted in the inability to connect to any ZigBee equipment like lights/sensors/switches.

This OS upgrade will fix problems with the Conbee 2 USB stick when running in a docker container.

As wmii seems no longer supported with Ubuntu 20.04 I switched the window manager to i3-vm as a replacement.