Add command line options for GCFFlasher (#376)
Allows the user to specify additional options to the flashing command that will be read by the script and checked against the possible options of the GCFFlasher_internal. Using this mechanism one could specify values for device-name, timeout, retries or the loglevel. Additionally the script defines a default of 60 seconds for the timeout-parameter.