April 2021 Update (#379)
* Add support for version-specific multiarch tag (#375)

* Add command line options for GCFFlasher (#376)

Allows the user to specify additional options to the flashing command that will be read by the script and checked against the possible options of the GCFFlasher_internal. Using this mechanism one could specify values for device-name, timeout, retries or the loglevel. Additionally the script defines a default of 60 seconds for the timeout-parameter.

* Improvements! (#377)

- Add support for noVNC (access VNC via a web browser)
- Add support for using Docker secrets or password file for VNC password
- Change image base to Ubuntu 20.04
- Change window manager from wmii to openbox

Co-authored-by: Jo-Be-Co <Dev@JoBeCo.de>
Co-authored-by: Philippe Delodder <7259755+phdelodder@users.noreply.github.com>